Include overridden CALICO_GIT_REF in the PR title
By default, we use either the branch name (for master) or the
major.minor (e.g. v3.29) of the branch name in our PR titles. If we
override CALICO_GIT_REF though, e.g. if we're updating for a tag, we
should be specific.

We now check the 'origin' of the variable in the Makefile; if it's
'file' then it's specified by the Makefile itself, therefore it's
the default value. Otherwise the variable was specified some other
how, e.g. 'command line' for passing it as an argument to make or
'environment' if it's exported to make from the shell, and so we respect
that ref.